Fix ansible-test imports and paths after refactor.
Also include a changelog entry covering this commit and the previous.
