+++ test/lib/ansible_test/_internal/commands/integration/cloud/gcp.py
@@ -4,12 +4,12 @@
 from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
-from ..util import (
+from ....util import (
     display,
     ConfigParser,
 )
 
-from ..config import (
+from ....config import (
     IntegrationConfig,
 )
