+++ test/lib/ansible_test/_internal/commands/integration/cloud/httptester.py
@@ -4,16 +4,16 @@ __metaclass__ = type
 
 import os
 
-from ..util import (
+from ....util import (
     display,
     generate_password,
 )
 
-from ..config import (
+from ....config import (
     IntegrationConfig,
 )
 
-from ..containers import (
+from ....containers import (
     run_support_container,
 )
