+++ test/lib/ansible_test/_internal/commands/integration/cloud/opennebula.py
@@ -2,7 +2,7 @@
 from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
-from ..util import (
+from ....util import (
     ConfigParser,
     display,
 )
