+++ test/lib/ansible_test/_internal/commands/integration/cloud/vcenter.py
@@ -4,17 +4,17 @@ __metaclass__ = type
 
 import os
 
-from ..util import (
+from ....util import (
     ApplicationError,
     ConfigParser,
     display,
 )
 
-from ..config import (
+from ....config import (
     IntegrationConfig,
 )
 
-from ..containers import (
+from ....containers import (
     run_support_container,
 )
