+++ test/lib/ansible_test/_internal/commands/integration/cloud/vultr.py
@@ -2,12 +2,12 @@
 from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
-from ..util import (
+from ....util import (
     ConfigParser,
     display,
 )
 
-from ..config import (
+from ....config import (
     IntegrationConfig,
 )
