+++ test/lib/ansible_test/_internal/commands/sanity/ansible_doc.py
@@ -6,41 +6,41 @@ import collections
 import os
 import re
 
-from .. import types as t
+from ... import types as t
 
-from ..sanity import (
+from . import (
     SanitySingleVersion,
     SanityFailure,
     SanitySuccess,
 )
 
-from ..target import (
+from ...target import (
     TestTarget,
 )
 
-from ..util import (
+from ...util import (
     SubprocessError,
     display,
     is_subdir,
 )
 
-from ..util_common import (
+from ...util_common import (
     intercept_command,
 )
 
-from ..ansible_util import (
+from ...ansible_util import (
     ansible_environment,
 )
 
-from ..config import (
+from ...config import (
     SanityConfig,
 )
 
-from ..data import (
+from ...data import (
     data_context,
 )
 
-from ..coverage_util import (
+from ...coverage_util import (
     coverage_context,
 )
