+++ test/lib/ansible_test/_internal/commands/sanity/pep8.py
@@ -4,9 +4,9 @@ __metaclass__ = type
 
 import os
 
-from .. import types as t
+from ... import types as t
 
-from ..sanity import (
+from . import (
     SanitySingleVersion,
     SanityMessage,
     SanityFailure,
@@ -14,11 +14,11 @@ from ..sanity import (
     SANITY_ROOT,
 )
 
-from ..target import (
+from ...target import (
     TestTarget,
 )
 
-from ..util import (
+from ...util import (
     SubprocessError,
     read_lines_without_comments,
     parse_to_list_of_dict,
@@ -26,11 +26,11 @@ from ..util import (
     is_subdir,
 )
 
-from ..util_common import (
+from ...util_common import (
     run_command,
 )
 
-from ..config import (
+from ...config import (
     SanityConfig,
 )
